% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{getCoverage}
\alias{getCoverage}
\title{caculate read coverage in the flanking region of target exons}
\usage{
getCoverage(bamfile, region, flanking.ws = 150, read.tag.names = F)
}
\arguments{
\item{bamfile}{- bam file of chip-seq data (bowtie alignment)}

\item{region}{- a dataframe indicate chromosome, target exon start and target exon end position}
}
\value{
list contain: 1. vector of read coverage of each individual position in the left flanking region
                      2. vector of read coverage of each individual position in the right flanking region
                      3. total library size of bam file
}
\description{
caculate read coverage in the flanking region of target exons
}
