% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.processing.R
\name{HMflankingSig}
\alias{HMflankingSig}
\title{This function is used to extract the hPTM signal in the exon flanking regions
or rMATS HM files and calssify alternative spliced exons into different
categories (splicing codes): gain (0), loss (1), High (2) and Low (3)}
\usage{
HMflankingSig(HM_file, marker, total_reads)
}
\arguments{
\item{HM_file:}{processed HM rMAST file with splicing categories}

\item{marker:}{hPTM information, e.g. H3K36me3}

\item{total_reads:}{total reads in sample}
}
\value{
file contains hPTM signal in the exon flanking regions
}
\description{
This function is used to extract the hPTM signal in the exon flanking regions
or rMATS HM files and calssify alternative spliced exons into different
categories (splicing codes): gain (0), loss (1), High (2) and Low (3)
}
