% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{iRFModel}
\alias{iRFModel}
\title{run iterative random forest model based on 5-fold CV}
\usage{
iRFModel(data, n.iter = 20, n.interaction = 10, bootstrap = 30)
}
\arguments{
\item{data}{training data}
}
\value{
interation list
}
\description{
run iterative random forest model based on 5-fold CV
}
