% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model}
\alias{model}
\title{run logistic and random forest model}
\usage{
model(data, seed = 11223)
}
\arguments{
\item{data}{training data}
}
\value{
list contains performance metric of logistic/random forest model and impscore
}
\description{
run logistic and random forest model
}
