% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{model_exp}
\alias{model_exp}
\title{random forest regression model to identify HMs asscociated with expression}
\usage{
model_exp(data)
}
\arguments{
\item{data}{training data}
}
\value{
list contains performance metric of logistic/random forest model and impscore
}
\description{
random forest regression model to identify HMs asscociated with expression
}
