% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{modelPerformance}
\alias{modelPerformance}
\title{Evaluate model performance}
\usage{
modelPerformance(pred, pred_prob, class, cat)
}
\arguments{
\item{pred}{predicted value}

\item{pred_prob}{predicted probility}

\item{class}{true class value}

\item{cat}{# of classes}
}
\value{
performance evaluation metrics
}
\description{
Evaluate model performance
}
