% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotCoverage}
\alias{plotCoverage}
\title{This function is used to visualize the chip-seq distribution of flanking regions}
\usage{
plotCoverage(
  HM_file,
  total_reads,
  sample_info,
  subsampleCanonical = F,
  CanonicalFile = NULL
)
}
\arguments{
\item{HM_file:}{files contain the chip-seq signals in the flanking region of AS exons and AS exon class}

\item{total_reads:}{number of total reads in the sample}

\item{sample_info:}{vector contains sample information (tissue, markers and time points)}

\item{subsampleCanonical:}{subsample signals from canonical exons}

\item{CanonicalFile:}{file contains chip-seq signal from canonical exons}
}
\value{
plot of distribution of hPTMs in the exon flanking regions
}
\description{
This function is used to visualize the chip-seq distribution of flanking regions
}
