% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotFeatureImp}
\alias{plotFeatureImp}
\title{This function is used to plot the top important markers based on exon flanking region position}
\usage{
plotFeatureImp(importantscore, title = "")
}
\arguments{
\item{importantscore}{impscore generated by model function}
}
\value{
plot of important markers
}
\description{
This function is used to plot the top important markers based on exon flanking region position
}
