% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotInteraction}
\alias{plotInteraction}
\title{plot interaction of histone markers}
\usage{
plotInteraction(interaction.list, cutoff = 0.5, tissue, timepoint)
}
\arguments{
\item{interaction.list:}{table contain interaction information}

\item{cutoff:}{cutoff value for visualization of interactions}

\item{tissue:}{tissue type}

\item{timepoint:}{data timepoint}
}
\value{
interaction plot
}
\description{
plot interaction of histone markers
}
