% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotModelPerformance}
\alias{plotModelPerformance}
\title{visualize the performance of random forest and logistic regression model}
\usage{
plotModelPerformance(rf_perf, logit_perf)
}
\arguments{
\item{rf_perf:}{performance table of random forest model}

\item{logit_perf:}{performance table of logistic regression}
}
\value{
list of bargraphs for model performance
}
\description{
visualize the performance of random forest and logistic regression model
}
