% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{plotRFperformance}
\alias{plotRFperformance}
\title{plot the performance of random forest model}
\usage{
plotRFperformance(rf_perf)
}
\arguments{
\item{rf_perf:}{performance table of random forest model}
}
\value{
lists of plots
}
\description{
plot the performance of random forest model
}
