% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{Visfeatures}
\alias{Visfeatures}
\title{This function is used to visulize the importance features for different gene expression categories}
\usage{
Visfeatures(importance.filelist, nTOP = 5)
}
\arguments{
\item{importance.filelist}{list of files contain important features based on gene expression categories (low, median, high)}

\item{nTOP}{visulize ntop features}
}
\value{
plot of top important feautres
}
\description{
This function is used to visulize the importance features for different gene expression categories
}
