% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis.R
\name{VisfeaturesMarks}
\alias{VisfeaturesMarks}
\title{This function is used to visulize the importance markers for different gene expression categories}
\usage{
VisfeaturesMarks(impscores, nTOP = 5)
}
\arguments{
\item{nTOP}{visulize ntop markers}

\item{importance.filelist}{list of files contain important features based on gene expression categories (low, median, high)}
}
\value{
plot of top important feautres
}
\description{
This function is used to visulize the importance markers for different gene expression categories
}
